Pull UCI helpers into separate crate
So we can consume it in other projects as well
