+++ chessmatch/src/engine.rs
@@ -4,8 +4,10 @@ use chess::board::Board;
 use tokio::process::{Command, ChildStdout};
 use tokio::io::{BufReader,  AsyncWriteExt, AsyncBufReadExt};
 use serde::Deserialize;
-
-use shared::uci::{UciClientMessage, UciEngineMessage,  SearchInfo, TCType};
+use uci::client::UciClientMessage;
+use uci::engine::UciEngineMessage;
+use uci::search_info::SearchInfo;
+use uci::time_control::TimeControl;
 
 #[derive(Debug, Clone, Deserialize)]
 pub struct EngineConfig {
@@ -22,7 +24,7 @@ pub struct Engine {
     stdin: tokio::process::ChildStdin,
     stdout: tokio::io::BufReader<ChildStdout>,
     pub config: EngineConfig,
-    pub tc: TCType,
+    pub tc: TimeControl,
     pub search_info: SearchInfo,
 }
 
@@ -42,13 +44,13 @@ impl Engine {
         let stdout = process.stdout.take().unwrap();
 
         let tc = if let Some(depth) = config.depth {
-            TCType::Depth(depth)
+            TimeControl::Depth(depth)
         } else if let Some(time) = config.time {
-            TCType::FixedTime(Duration::from_millis(time as u64))
+            TimeControl::FixedTime(Duration::from_millis(time as u64))
         } else if let Some(nodes) = config.nodes {
-            TCType::Nodes(nodes)
+            TimeControl::Nodes(nodes)
         } else {
-            TCType::Infinite
+            TimeControl::Infinite
         };
 
         Self {
