+++ shared/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod uci;
 pub mod components;
