+++ simbelmyne/Cargo.toml
@@ -7,9 +7,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = "1.0.75"
 chess = { path = "../chess" }
+uci = { path = "../uci" }
 shared = { path = "../shared" }
+anyhow = "1.0.75"
 colored = "2.0.4"
 clap = { version = "4.4.7", features = ["derive"] }
 crossterm = "0.27.0"
