+++ simbelmyne/src/cli/bench.rs
@@ -1,5 +1,5 @@
 use colored::Colorize;
-use shared::uci::TCType;
+use uci::time_control::TimeControl;
 
 use crate::{position::Position, transpositions::TTable, time_control::TimeController};
 
@@ -18,7 +18,7 @@ pub fn run_single(fen: &str, depth: usize) {
     let board = fen.parse().unwrap();
     let position = Position::new(board);
     let mut tt = TTable::with_capacity(64);
-    let (tc, _handle) = TimeController::new(TCType::Depth(depth), board.current);
+    let (tc, _handle) = TimeController::new(TimeControl::Depth(depth), board.current);
     let search = position.search(&mut tt, tc);
 
     println!("{board}");
