+++ simbelmyne/src/main.rs
@@ -24,6 +24,17 @@ struct Cli {
 fn main() -> anyhow::Result<()> {
     let cli = Cli::parse();
 
+    println!(r"
+ ,-.            .       .                   
+(   `           |       |                   
+ `-.  . . ;-.-. |-. ,-. | ;-.-. . . ;-. ,-. 
+.   ) | | | | | | | |-' | | | | | | | | |-' 
+ `-'  `-| ' ' ' `-' `-' ' ' ' ' `-| ' ' `-' 
+      `-'                       `-'         
+
+                                Version 1.0
+");
+
     if let Some(command) = cli.command {
         command.run()?;
     }  else {
