+++ uci/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "uci"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.75"
+chess = { path = "../chess" }
