+++ uci/src/lib.rs
@@ -1,0 +1,7 @@
+//! Utilities for serializing and deserializing Simbelmyne data into
+//! UCI messages
+
+pub mod client;
+pub mod engine;
+pub mod time_control;
+pub mod search_info;
